requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alabaster==0.7.16
altair==5.5.0
anamnesis==1.0.4
annotated-types==0.7.0
anyio==4.8.0
apptools==5.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
autoreject==0.4.3
babel==2.16.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
bokeh==3.6.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
darkdetect==0.8.0
dask==2025.1.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distributed==2025.1.0
docutils==0.21.2
edfio==0.4.6
eeglabio==0.0.3
envisage==7.0.3
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastjsonschema==2.21.1
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fslpy==3.21.1
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
glmtools==0.2.1
gmpy2==2.1.5
google-pasta==0.2.0
grpcio==1.70.0
h2==4.1.0
h5io==0.2.4
h5py==3.12.1
hpack==4.0.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
imagesize==1.4.1
importlib-metadata==4.13.0
importlib_resources==6.4.5
isoduration==20.11.0
Jinja2==3.0.3
joblib==1.4.2
json-tricks==3.17.3
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter_server==2.15.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
keras==3.8.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
llvmlite==0.44.0rc2
locket==1.0.0
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mayavi
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.4.1
mne==1.8.0
mne-bids==0.16.0
mne-bids-pipeline==1.9.0
mne-icalabel==0.7.0
mne-qt-browser==0.6.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munkres==1.1.4
namex==0.0.8
narwhals==1.22.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neurokit2==0.2.10
nibabel==5.3.2
nilearn==0.11.1
numba
numexpr==2.10.2
numpy==1.26.4
numpydoc==1.8.0
onnxruntime==1.20.1
#opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
panel==1.5.5
param==2.2.0
parse==1.20.2
partd==1.4.2
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
ply==3.11
pooch==1.8.2
prometheus_client==0.21.1
propcache==0.3.0
protobuf==5.28.3
psutil==6.1.1
ptyprocess==0.7.0
pyarrow==19.0.0
pybv==0.7.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyface==8.0.0
Pygments==2.19.1
pymatreader==1.0.0
PyOpenGL==3.1.6
pyparsing==3.2.1
PyQt5==5.12.3
PyQt5_sip
PyQtChart==5.12
pyqtgraph==0.12.4
PyQtWebEngine==5.12.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-picard==0.8
pytz==2024.2
pyvista==0.38.5
pyvistaqt==0.11.1
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.2.1
QDarkStyle==3.2.3
QtPy==2.4.2
referencing==0.36.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
sails==1.6.0
scikit-learn==1.6.1
scipy==1.15.1
scooby==0.10.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.6.0
simpervisor
sip
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==7.3.7
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stpyvista==0.1.4
streamlit==1.41.1
sympy==1.13.3
tabulate==0.9.0
tblib==3.0.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unicodedata2==16.0.0
uri-template==1.3.0
urllib3==2.3.0
vtk==9.2.2
watchdog==6.0.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xmltodict==0.14.2
xyzservices==2024.9.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
zstandard==0.23.0
streamlit-ace==0.1.1
pybids==0.19.0
sleepecg==0.5.5
autoreject==0.4.3
redis==5.2.1
vtk-osmesa==9.0.1
